Further thoughts on FilterHotPixel.
After performing a toy study, I found that for a threshold of 0.01
and with 60 pixel hits in each unique event, 600 unique events must be considered
to accurately find hot pixels, and not accidently call pixels 'hot' due to
statistical fluctuations. See plot.

As we have scaler 15, this means 10,000 data events sould be used as a minimum
with threshold 0.01.

I added this over-ride in FilterHotPixel, and also set the threshold to 0.01 in
pyEudetReconstructionOnly.py.

Note, if the threshold is lowered then even more events must be used to determine
hot pixels.