Altered FindMatchedCluster to consider track-pixel distances instead of track-cluster distances.
This is important because for large clusters (possibly delta rays) and for angled tracks,
the track position could be far from the cluster centre, but should still be close to
one of the hit pixels in the cluster.

In Cluster.py:
- new function GetPixelResiduals to get the minimum track-pixel distance for a cluster

In EudetData.py:
- alteration of FindMatchedCluster to use this new function

As a result of this change, more clusters are matched to tracks:
146264 (pixel matching) vs 146064 (cluster matching)
and the distance histogram is sharper.