Added CountPixelMapRepeats function.
In ToolBox.py:
- added new function CountPixelMapRepeats. This counts how often
the pixel map is repeated before there is a new event. It is the
old 'scaler' variable, which was thought to be 15 but is shown to
be never more than 15 but often less. This shows that we shouldn't
use a 'scaler' variable, but rather check for when there's a new
pixel map by hand.

In pyEudetReconstructionOnly.py:
- call this function and save histogram