Replaced EtaCorrection method for s2x2y2 clusters with maxTOT method.
In Cluster.py:
- replaced the EtaCorrection method with the maxTOT method for s2x2y2 clusters.
These are two hit clusters in which the pixels are corner to corner. I think
most of them are not caused by charge sharing, but rather noise / cross talk /
delta rays.

Taking the hottest pixel as the centre improves the fitted resolution of these
events, see plots. However, there are more outliers, so the RMS remains about
the same. Nevertheless, the overall resolution with all cluster sizes does decrease
slightly with this change.

To do in the future:
- it might be possible to improve further by considering the Qrel of the cluster,
and performing maxTOT for some range and EtaCorrection for the rest.