Added resolution plots to ComputeAlignment to better see quality of alignment.
In ComputeAlignment.py:
- added total and 2hit cluster resolutin plots to the end, to be able to see the
quality of the alignment as soon as it's finished, without having to run
reconstruction and analysis seperately. Plots are saved in rootfile. Results are
printed at the end of the alignment process.

Note: I realise that this is alot of 'code' in what should just be a user script.
In the future it might be good to hide the code in a function, but currently no
MakeResolutionPlots exists.... in pyEudetAnalysisOnly.py the plots are also made
here instead of being resuls of a function.